+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+- Allow a custom expiration time for custom tokens. 
+
 ## 1.5.0 - 2017-04-03
 
 - Allow the usage of a custom key store when using the Handler.